+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Unreleased
 
+## 0.14.1 - 2026-05-11
+
 ### Added
 - resource `storage_zone`: S3 support;
 - resource `compute_container_app`: container `image_digest` attribute;